refactor removing component
